pyproject.toml
[project]
name = "exscribe"
version = "0.0.1"
description = "Extract and transcribe subtitles from videos using multimodal models."
keywords = ["OCR", "subtitles", "video", "transcription", "multimodal", "LLM", "OpenAI"]
readme = "README.md"
authors = [{ name = "m-krastev", email = "matey.krastev@outlook.com" }]
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "av",
    "opencv-python",
    "google-generativeai",
    "numpy",
    "Pillow",
    "scikit-image",
    "tqdm",
    "google-genai",
]

[project.urls]
homepage = "https://github.com/m-krastev/exscribe"
issues = "https://github.com/m-krastev/exscribe/issues"

[project.scripts]
exscribe = "exscribe:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
